Both of the Extra sections are now Functional

+++ src/Pages/MyList/MyList.jsx
@@ -20,28 +20,39 @@ const MyList = () => {
 
     const handleDelete = id => {
 
-        // Swal.fire({
-        //     title: "Are you sure?",
-        //     text: "You won't be able to revert this!",
-        //     icon: "warning",
-        //     showCancelButton: true,
-        //     confirmButtonColor: "#3085d6",
-        //     cancelButtonColor: "#d33",
-        //     confirmButtonText: "Yes, delete it!"
-        //   }).then((result) => {
-        //     if (result.isConfirmed) {
-        //       Swal.fire({
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+          }).then((result) => {
+            if (result.isConfirmed) {
+            fetch(`https://discover-euro-server-r9x6oq445-s-m-saifur-rahmans-projects.vercel.app/spots/${id}`, {
+                method:'DELETE'
+               })
+               .then(res => res.json())
+               .then(data =>{
+                         Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success"
+              });
+               })
+
+            }
+          });
+
+         //       Swal.fire({
         //         title: "Deleted!",
         //         text: "Your file has been deleted.",
         //         icon: "success"
         //       });
-        //     }
-        //   });
 
 
-       fetch(`https://discover-euro-server-r9x6oq445-s-m-saifur-rahmans-projects.vercel.app/spots/${id}`, {
-        method:'DELETE'
-       })
+       
     //    .then(res=>{
     //     console.log(res);
     //    })
